fix: not found jobs to return 404 instead 500
